Voice_Website/requirements.txt
atomicwrites==1.4.1
attrs==23.1.0
azure-cognitiveservices-speech==1.31.0
blinker==1.6.2
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
distlib==0.3.6
ffmpeg-python==0.2.0
filelock==3.10.0
Flask==3.0.0
Flask-Session==0.5.0
future==0.18.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
llvmlite==0.40.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.2.1
networkx==3.0
numba==0.57.1
numpy==1.24.1
openai-whisper==20230314
packaging==23.1
Pillow==9.3.0
platformdirs==3.1.1
pluggy==0.13.1
py==1.11.0
PyAudio==0.2.13
pytest==6.2.4
regex==2023.8.8
requests==2.31.0
SpeechRecognition==3.8.1
sympy==1.11.1
tiktoken==0.3.1
toml==0.10.2
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tqdm==4.66.1
typing_extensions==4.4.0
urllib3==2.0.4
virtualenv==20.21.0
websockets==11.0.3
Werkzeug==3.0.0
zipp==3.16.2